pyproject.toml
[project]
name = "anyon-tech-spec-agent"
version = "1.0.0"
description = "Technical Specification Agent for ANYON Platform - Automated TRD Generation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ANYON Team", email = "team@anyon.platform"}
]

dependencies = [
    # Web Framework
    "fastapi>=0.121.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",

    # LangGraph & LangChain
    "langgraph>=1.0.3",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.4.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.0.20",

    # LLM APIs
    "anthropic>=0.40.0",
    "openai>=1.58.0",

    # Database
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",

    # Caching & Queues
    "redis>=5.2.0",
    "hiredis>=2.3.0",

    # HTTP & Async
    "httpx>=0.28.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",

    # Web Scraping & Search
    "beautifulsoup4>=4.12.3",
    "tavily-python>=0.5.0",

    # Code Parsing
    "tree-sitter>=0.22.0",
    "tree-sitter-typescript>=0.21.0",

    # Diagram Generation
    "diagrams>=0.24.0",

    # Configuration & Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",

    # Logging & Monitoring
    "structlog>=24.4.0",
    "prometheus-client>=0.20.0",

    # Security
    "pyjwt[crypto]>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",

    # Utilities
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-timeout>=2.3.0",

    # Type Checking
    "mypy>=1.13.0",
    "types-redis>=4.6.0",
    "types-aiofiles>=23.2.0",

    # Code Quality
    "ruff>=0.8.0",
    "black>=24.10.0",
    "isort>=5.13.0",

    # Development Tools
    "ipython>=8.20.0",
    "ipdb>=0.13.13",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
